Refactor AI analysis logic and update interfaces for improved data handling
